ema_backend/scripts/requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.1
fuzzywuzzy==0.18.0
h11==0.14.0
idna==3.7
JPype1==1.5.0
lxml==5.2.2
mysql-connector-python==8.4.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tabula-py==2.9.3
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wsproto==1.2.0
